Refactor to improve maintainability
