+++ src/PSRule/Data/InputFileInfoCollection.cs
@@ -8,19 +8,6 @@ namespace PSRule.Data;
 /// <summary>
 /// A collection of <see cref="InputFileInfo"/>.
 /// </summary>
-public interface IInputFileInfoCollection : IEnumerable<InputFileInfo>
-{
-    /// <summary>
-    /// Filters the collection to only include <see cref="InputFileInfo"/> with a specific file extension.
-    /// </summary>
-    /// <param name="extension">A file extension to filter the collection to.</param>
-    /// <returns>A filtered collection.</returns>
-    IInputFileInfoCollection WithExtension(string extension);
-}
-
-/// <summary>
-/// A collection of <see cref="InputFileInfo"/>.
-/// </summary>
 internal sealed class InputFileInfoCollection : IInputFileInfoCollection, IEnumerable<InputFileInfo>
 {
     private readonly IEnumerable<InputFileInfo> _Items;
