+++ src/PSRule/Definitions/SuppressionGroups/ISuppressionGroupV1Spec.cs
@@ -1,8 +1,7 @@
-// Copyright (c) Microsoft Corporation.
+﻿// Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
 using PSRule.Definitions.Expressions;
-using PSRule.Pipeline;
 
 namespace PSRule.Definitions.SuppressionGroups;
 
@@ -28,29 +27,4 @@ internal interface ISuppressionGroupV1Spec
     /// An expression. If the expression evaluates as true and rules specified by <see cref="Rule"/> are suppressed.
     /// </summary>
     LanguageIf If { get; }
-}
-
-/// <summary>
-/// A suppression group resource V1.
-/// </summary>
-[Spec(Specs.V1, Specs.SuppressionGroup)]
-internal sealed class SuppressionGroupV1 : InternalResource<SuppressionGroupV1Spec>
-{
-    public SuppressionGroupV1(string apiVersion, SourceFile source, ResourceMetadata metadata, IResourceHelpInfo info, ISourceExtent extent, SuppressionGroupV1Spec spec)
-        : base(ResourceKind.SuppressionGroup, apiVersion, source, metadata, info, extent, spec) { }
-}
-
-/// <summary>
-/// A specification for a V1 suppression group resource.
-/// </summary>
-internal sealed class SuppressionGroupV1Spec : Spec, ISuppressionGroupV1Spec
-{
-    /// <inheritdoc/>
-    public DateTime? ExpiresOn { get; set; }
-
-    /// <inheritdoc/>
-    public string[] Rule { get; set; }
-
-    /// <inheritdoc/>
-    public LanguageIf If { get; set; }
 }
