+++ src/PSRule/Pipeline/Formatters/AzurePipelinesFormatter.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT License.
 
 using PSRule.Configuration;
+using PSRule.Definitions.Rules;
 using PSRule.Rules;
 
 namespace PSRule.Pipeline.Formatters;
@@ -50,13 +51,13 @@ internal sealed class AzurePipelinesFormatter : AssertFormatterBase, IAssertForm
     {
         base.FailDetail(record);
         var message = GetFailMessage(record);
-        if (record.Level == Definitions.SeverityLevel.Error)
+        if (record.Level == SeverityLevel.Error)
             Error(message);
 
-        if (record.Level == Definitions.SeverityLevel.Warning)
+        if (record.Level == SeverityLevel.Warning)
             Warning(message);
 
-        if (record.Level != Definitions.SeverityLevel.Information)
+        if (record.Level != SeverityLevel.Information)
             LineBreak();
     }
 }
